+++ strprofiler/strprofiler.py
